% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{prettytext}
\alias{prettytext}
\title{Make strings prettier for printing}
\usage{
prettytext(
  string,
  capitalisation = c("title", "all", "first", "lower", "unchanged"),
  trimws = TRUE,
  space_camel = TRUE,
  space_snake = TRUE,
  space_bracket = TRUE
)
}
\arguments{
\item{string}{A character vector, or something coercible to one.}

\item{capitalisation}{Which words should be capitalised (see Details for more info)}

\item{trimws}{trim leading and trailing and whitespace}

\item{space_camel}{turn camelCase into camel space (flag)}

\item{space_snake}{turn snake_case into snake space  (flag)}

\item{space_bracket}{turn bracketed(words) into bracketed (words)  (flag)}
}
\value{
human friendly formatted version of input string (character)
}
\description{
Takes an input string and 'beautify' it by converting from 'programmerStyle' to 'Human Friendly Formatted'
}
\details{
\strong{Capitalisation Options}\tabular{ll}{
   foo \tab bar \cr
   title \tab This is Title Case \cr
   all \tab Every Word Is Capitalised \cr
   first \tab First word is capitalised \cr
   lowercase \tab every letter is lower case \cr
   unchanged \tab Whatever is INPUT IS OUTPUT \cr
}
}
